暂时屏蔽备份功能，修复错误的使用auto导致数据拷贝的BUG

+++ src/tt_client.cc
@@ -47,7 +47,7 @@ namespace tokyotyrant {
             tcp_client_->ConnectTo(addr, true);
             state_ = ConnectionState::kConnecting;
             co_->Yield();
-            return Connected() ? kOk : kRefused;
+            return Connected() ? static_cast<int>(kOk) : static_cast<int>(kRefused);
         }
     }
